+++ demos/browser/index.html
@@ -41,7 +41,7 @@
 	<script src="../../dist/pptxgen.min.js"></script>
 	-->
 	<script src="./js/pptxgen.bundle.js"></script>
-	<script src="../modules/demos.js" type="module"></script>
+	<script src="../modules/demos.mjs" type="module"></script>
 
 	<script type="module">
 		import { doAppStart, execGenSlidesFunc, runAllDemos, table2slides1, table2slides2, table2slidesDemoForTab } from './js/browser.js';
