+++ demos/modules/demo_text.mjs
@@ -7,7 +7,7 @@
  * BLD.: 20210401
  */
 
-import { BASE_TABLE_OPTS, BASE_TEXT_OPTS_L, BASE_TEXT_OPTS_R, LOREM_IPSUM_ENG } from "./enums.js";
+import { BASE_TABLE_OPTS, BASE_TEXT_OPTS_L, BASE_TEXT_OPTS_R, LOREM_IPSUM_ENG } from "./enums.mjs";
 
 export function genSlides_Text(pptx) {
 	pptx.addSection({ title: "Text" });
