+++ demos/node/demo.js
@@ -1,7 +1,7 @@
 /*
  * NAME: demo.js
  * AUTH: Brent Ely (https://github.com/gitbrent/)
- * DATE: 20200516
+ * DATE: 20210403
  * DESC: PptxGenJS feature demos for Node.js
  * REQS: npm 4.x + `npm install pptxgenjs`
  *
@@ -11,8 +11,9 @@
  */
 
 // ============================================================================
+import { execGenSlidesFuncs, runEveryTest } from "../modules/demos.mjs";
 let PptxGenJS = require("pptxgenjs");
-let demo = require("../common/demos.js");
+//let demo = require("../modules/demos.js");
 let pptx = new PptxGenJS();
 let exportName = "PptxGenJS_Demo_Node";
 
@@ -24,8 +25,8 @@ if (process.argv.length > 2) {
 	// A: Run predefined test from `../common/demos.js` //-OR-// Local Tests (callbacks, etc.)
 	Promise.resolve()
 		.then(() => {
-			if (process.argv[2].toLowerCase() === "all") return demo.runEveryTest();
-			return demo.execGenSlidesFuncs(process.argv[2]);
+			if (process.argv[2].toLowerCase() === "all") return runEveryTest();
+			return execGenSlidesFuncs(process.argv[2]);
 		})
 		.catch((err) => {
 			throw new Error(err);
